refactor: move `source_order_name_lookup` to `parse.h`
